Remove cast

+++ utility/CalculateInstalledMemory.cpp
@@ -35,7 +35,7 @@ namespace system {
     std::uint64_t num_pages = static_cast<std::uint64_t>(sysconf(_SC_PHYS_PAGES));
     std::uint64_t page_size = static_cast<std::uint64_t>(sysconf(_SC_PAGE_SIZE));
     if (num_pages > 0 &&  page_size > 0) {
-      total_memory = static_cast<std::uint64_t>(num_pages * page_size);
+      total_memory = num_pages * page_size;
       LOG(INFO) << "Total memory is " << total_memory << " bytes\n";
       return true;
     }
